requirements.txt
aiogram
python-dotenv
openai
Pillow
